# backend/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package openapi

import (
	"io"
	"net/url"

	"github.com/google/uuid"
)

// Ref: #/components/schemas/APIError
type APIError struct {
	Code      string `json:"code"`
	Message   string `json:"message"`
	RequestID string `json:"request_id"`
}

// GetCode returns the value of Code.
func (s *APIError) GetCode() string {
	return s.Code
}

// GetMessage returns the value of Message.
func (s *APIError) GetMessage() string {
	return s.Message
}

// GetRequestID returns the value of RequestID.
func (s *APIError) GetRequestID() string {
	return s.RequestID
}

// SetCode sets the value of Code.
func (s *APIError) SetCode(val string) {
	s.Code = val
}

// SetMessage sets the value of Message.
func (s *APIError) SetMessage(val string) {
	s.Message = val
}

// SetRequestID sets the value of RequestID.
func (s *APIError) SetRequestID(val string) {
	s.RequestID = val
}

// Ref: #/components/schemas/ErrorEnvelope
type ErrorEnvelope struct {
	Error APIError `json:"error"`
}

// GetError returns the value of Error.
func (s *ErrorEnvelope) GetError() APIError {
	return s.Error
}

// SetError sets the value of Error.
func (s *ErrorEnvelope) SetError(val APIError) {
	s.Error = val
}

func (*ErrorEnvelope) listVideosRes() {}

type GetVideoBadRequest ErrorEnvelope

func (*GetVideoBadRequest) getVideoRes() {}

type GetVideoInternalServerError ErrorEnvelope

func (*GetVideoInternalServerError) getVideoRes() {}

type GetVideoNotFound ErrorEnvelope

func (*GetVideoNotFound) getVideoRes() {}

type HealthzOK struct {
	Data io.Reader
}

// Read reads data from the Data reader.
//
// Kept to satisfy the io.Reader interface.
func (s HealthzOK) Read(p []byte) (n int, err error) {
	if s.Data == nil {
		return 0, io.EOF
	}
	return s.Data.Read(p)
}

type ListVideosOKApplicationJSON []Video

func (*ListVideosOKApplicationJSON) listVideosRes() {}

type ReadyzOK struct {
	Data io.Reader
}

// Read reads data from the Data reader.
//
// Kept to satisfy the io.Reader interface.
func (s ReadyzOK) Read(p []byte) (n int, err error) {
	if s.Data == nil {
		return 0, io.EOF
	}
	return s.Data.Read(p)
}

func (*ReadyzOK) readyzRes() {}

type ReadyzServiceUnavailable struct {
	Data io.Reader
}

// Read reads data from the Data reader.
//
// Kept to satisfy the io.Reader interface.
func (s ReadyzServiceUnavailable) Read(p []byte) (n int, err error) {
	if s.Data == nil {
		return 0, io.EOF
	}
	return s.Data.Read(p)
}

func (*ReadyzServiceUnavailable) readyzRes() {}

type RecordVideoViewBadRequest ErrorEnvelope

func (*RecordVideoViewBadRequest) recordVideoViewRes() {}

type RecordVideoViewInternalServerError ErrorEnvelope

func (*RecordVideoViewInternalServerError) recordVideoViewRes() {}

type RecordVideoViewNotFound ErrorEnvelope

func (*RecordVideoViewNotFound) recordVideoViewRes() {}

// Ref: #/components/schemas/RecordViewRequest
type RecordViewRequest struct {
	ViewSessionID string `json:"view_session_id"`
}

// GetViewSessionID returns the value of ViewSessionID.
func (s *RecordViewRequest) GetViewSessionID() string {
	return s.ViewSessionID
}

// SetViewSessionID sets the value of ViewSessionID.
func (s *RecordViewRequest) SetViewSessionID(val string) {
	s.ViewSessionID = val
}

// Ref: #/components/schemas/RecordViewResponse
type RecordViewResponse struct {
	VideoID uuid.UUID `json:"video_id"`
	Views   int64     `json:"views"`
	Counted bool      `json:"counted"`
}

// GetVideoID returns the value of VideoID.
func (s *RecordViewResponse) GetVideoID() uuid.UUID {
	return s.VideoID
}

// GetViews returns the value of Views.
func (s *RecordViewResponse) GetViews() int64 {
	return s.Views
}

// GetCounted returns the value of Counted.
func (s *RecordViewResponse) GetCounted() bool {
	return s.Counted
}

// SetVideoID sets the value of VideoID.
func (s *RecordViewResponse) SetVideoID(val uuid.UUID) {
	s.VideoID = val
}

// SetViews sets the value of Views.
func (s *RecordViewResponse) SetViews(val int64) {
	s.Views = val
}

// SetCounted sets the value of Counted.
func (s *RecordViewResponse) SetCounted(val bool) {
	s.Counted = val
}

func (*RecordViewResponse) recordVideoViewRes() {}

// Ref: #/components/schemas/Video
type Video struct {
	ID      uuid.UUID `json:"id"`
	Title   string    `json:"title"`
	M3u8URL url.URL   `json:"m3u8_url"`
	Views   int64     `json:"views"`
}

// GetID returns the value of ID.
func (s *Video) GetID() uuid.UUID {
	return s.ID
}

// GetTitle returns the value of Title.
func (s *Video) GetTitle() string {
	return s.Title
}

// GetM3u8URL returns the value of M3u8URL.
func (s *Video) GetM3u8URL() url.URL {
	return s.M3u8URL
}

// GetViews returns the value of Views.
func (s *Video) GetViews() int64 {
	return s.Views
}

// SetID sets the value of ID.
func (s *Video) SetID(val uuid.UUID) {
	s.ID = val
}

// SetTitle sets the value of Title.
func (s *Video) SetTitle(val string) {
	s.Title = val
}

// SetM3u8URL sets the value of M3u8URL.
func (s *Video) SetM3u8URL(val url.URL) {
	s.M3u8URL = val
}

// SetViews sets the value of Views.
func (s *Video) SetViews(val int64) {
	s.Views = val
}

func (*Video) getVideoRes() {}