# backend/internal/repo/postgres/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Video struct {
	ID         pgtype.UUID        `db:"id" json:"id"`
	Title      string             `db:"title" json:"title"`
	M3u8Url    string             `db:"m3u8_url" json:"m3u8_url"`
	ViewsCount int64              `db:"views_count" json:"views_count"`
	CreatedAt  pgtype.Timestamptz `db:"created_at" json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `db:"updated_at" json:"updated_at"`
}

type VideoViewEvent struct {
	ID            pgtype.UUID        `db:"id" json:"id"`
	VideoID       pgtype.UUID        `db:"video_id" json:"video_id"`
	ViewSessionID string             `db:"view_session_id" json:"view_session_id"`
	CreatedAt     pgtype.Timestamptz `db:"created_at" json:"created_at"`
}